refactor: move auth wrapper path
